GWimbichCycles, TallFire&FloorSpikes, PyrosPickup
Added Great Wimbich Cycles, Tall Fire & Floor Spikes environmental enemies and the script for spawning the Pyros pickup after the Flute is played. The script for spawning the flames and the pyros pickup and killing the Rope (snake) enemy is currently located in the object Entity_Player but should be moved to the Rope when it is implemented.
Still need to implement using the pyros spell too